simplified loading.js and removed clear-loading.js
