+++ index.html
@@ -95,9 +95,7 @@
   <script type="module" src="/js/components/_swiper-card.js"></script>
   <script type="module" src="/js/components/_modal.js"></script>
   <script type="module" src="/js/render/shared/loading.js"></script>
-  <script type="module" src="/js/render/shared/clear-loading.js"></script>
   <script type="module" src="/js/render/shared/error.js"></script>
-  <script type="module" src="/js/render/shared/clear-error.js"></script>
   <script type="module" src="/js/render/shared/copyright.js"></script>
   <script type="module" src="/js/render/home-page/movies.js"></script>
   <script type="module" src="/js/render/home-page/movie-list-page.js"></script>
