+++ js/render/movie-details-page/movie-page.js
@@ -1,8 +1,7 @@
-import {renderLoading} from "../shared/loading.js";
+import {renderLoading, clearLoading} from "../shared/loading.js";
 import {renderCopyright} from "../shared/copyright.js";
 import {renderMovie} from "./movie.js";
 import {renderRelatedMovies} from "./related-movies.js";
-import {clearLoading} from "../shared/clear-loading.js";
 
 async function initMoviePage() {
     renderLoading();
