--- js/render/shared/clear-error.js
@@ -1,6 +1,0 @@
-function clearError() {
-    const $error = document.querySelector("#error");
-    $error.remove();
-}
-
-export {clearError};
