--- js/render/shared/clear-loading.js
@@ -1,6 +1,0 @@
-function clearLoading() {
-    const $loading = document.querySelector("#loading");
-    $loading.remove();
-  }
-
-export {clearLoading};
