+++ movie.html
@@ -115,9 +115,7 @@
   <script type="module" src="/js/components/_no-related-titles.js"></script>
   <script type="module" src="/js/components/_related-movies.js"></script>
   <script type="module" src="/js/render/shared/loading.js"></script>
-  <script type="module" src="/js/render/shared/clear-loading.js"></script>
   <script type="module" src="/js/render/shared/error.js"></script>
-  <script type="module" src="/js/render/shared/clear-error.js"></script>
   <script type="module" src="/js/render/shared/copyright.js"></script>
   <script type="module" src="/js/render/shared/swiper.js"></script>
   <script type="module" src="/js/render/movie-details-page/movie.js"></script>
